requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
decorator==4.4.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.1.2
packaging==24.1
pefile==2023.2.7
pillow==11.0.0
proglog==0.1.10
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.3.0
tqdm==4.66.6
urllib3==2.2.3
